requirements.txt
streamlit
python-dotenv
chromadb
openai

